add comments

+++ demo/demo/Program.cs
@@ -22,6 +22,13 @@ namespace demo
       System.Windows.Forms.Application.Run(new Form1());
     }
 
+    /// <summary>
+    /// Adds the items in the mapsWeb using the URL to the database.
+    /// Generates the Excel report using the user input.
+    /// </summary>
+    /// <param name="input">User input on what to query</param>
+    /// <param name="label">label in the UI to notify user if it was successful/failure</param>
+    /// <param name="mapsWeb">Map of product names and their URLs</param>
     public void GenerateReport(String input, System.Windows.Forms.Label label, Dictionary<String, String> mapsWeb)
     {
       ExcelActions xla = new ExcelActions();
@@ -29,20 +36,30 @@ namespace demo
       SqlConnection connection = null;
       try
       {
-        String excelOut = Directory.GetCurrentDirectory() + "\\out.xlsx";
-        //start to create the excel sheet.
+        if (String.IsNullOrEmpty(input))
+        {
+          Error(label, "User input is empty", xla, sh, connection);
+          return;
+        }
+
+        String excelOut = $@"{Directory.GetCurrentDirectory()}\Sheet1.xlsx";
+        //remove existing excel sheet.
         if (File.Exists(excelOut))
           File.Delete(excelOut);
 
+        //Add the products to DB
         AddToDB(mapsWeb, label, sh, ref connection);
 
+        //Select the columns where it matches the user input
         String userOutput = sh.ConditionalSelectAllTable(connection, "products", "type='" + input + "'");
 
         if (String.IsNullOrEmpty(userOutput))
         {
-          throw new Exception("");
+          Error(label, "Database is empty", xla, sh, connection);
+          return;
         }
 
+        //create and write the name, price and review to the excel sheet
         xla.GetWorkbook(excelOut);
         xla.QuickSave(excelOut);
         _Worksheet ws = xla.SetWorksheetName(excelOut, "sample");
@@ -50,39 +67,51 @@ namespace demo
         xla.WriteToXL(ws, 1, 2, "Review");
         xla.WriteToXL(ws, 1, 3, "Price");
 
+        //parse the sql db output to add the values to excel
         string[] userOutputArray = userOutput.Split('\n');
         for (int row = 0; row < userOutputArray.Length; row++)
         {
           string[] items = userOutputArray[row].Split('|');
-          for (int col = 0; col < items.Length-1; col++)
+          for (int col = 0; col < items.Length - 1; col++)
           {
-            xla.WriteToXL(ws, row+2, col+1, items[col]);
+            xla.WriteToXL(ws, row + 2, col + 1, items[col]);
           }
         }
 
-        sh.DeleteTable(connection, "products");
-        sh.CloseSQLConnection(conn: ref connection);
+        //auto arrange the rows
         dynamic allDataRange = ws.UsedRange;
         allDataRange.WrapText = false;
         allDataRange.Columns.AutoFit();
         allDataRange.Sort(Key1: allDataRange.Columns[1], Order1: XlSortOrder.xlAscending, Header: XlYesNoGuess.xlYes);
         ws.ListObjects.AddEx(XlListObjectSourceType.xlSrcRange, Type.Missing, Type.Missing, XlYesNoGuess.xlYes, Type.Missing, Type.Missing);
         xla.SaveWorkbook(excelOut);
-        xla.CloseXL();
 
+        //show to user that sheet is created
         label.Visible = true;
         label.BackColor = System.Drawing.Color.Yellow;
         label.ForeColor = System.Drawing.Color.Green;
         label.Text = "Generated Excel in Documents Folder";
-      }
-      catch (Exception e)
-      {
+
+        //close the excel sheet
         xla.CloseXL();
-        Console.WriteLine(e);
-        Error(label, "Unable to create Excel");
+        //delete the sql table
+        sh.DeleteTable(connection, "products");
+        //close sql connection
+        sh.CloseSQLConnection(conn: ref connection);
+      }
+      catch
+      {
+        //update the label to notiify to user there is an error
+        Error(label, "Unable to create Excel", xla, sh, connection);
+        return;
       }
     }
 
+    /// <summary>
+    /// Create a process to call python script and using URL to generate XML
+    /// </summary>
+    /// <param name="input">String URL of the webpage to scrape</param>
+    /// <param name="process">Process passed down to generate DB (console call)</param>
     public void WebProcess(String input, ref System.Diagnostics.Process process)
     {
       System.Diagnostics.ProcessStartInfo startInfo = new System.Diagnostics.ProcessStartInfo();
@@ -92,22 +121,46 @@ namespace demo
       process.Start();
     }
 
-    public void Error(System.Windows.Forms.Label label, String text)
+    /// <summary>
+    /// Communicates to the UI and closes opened Excel Sheet and DB connection if error occurs.
+    /// </summary>
+    /// <param name="label">label in the UI to notify user if it was successful/failure</param>
+    /// <param name="text">String text to display to UI</param>
+    /// <param name="xla">ExcelActions Object inorder to close</param>
+    /// <param name="sh">SQLHandler Object inorder to close</param>
+    /// <param name="connection">to delete the table and disconnect</param>
+    public void Error(System.Windows.Forms.Label label, String text, ExcelActions xla, SQLHandler sh, SqlConnection connection)
     {
       label.Visible = true;
       label.BackColor = System.Drawing.Color.Yellow;
       label.ForeColor = System.Drawing.Color.Red;
       label.Text = text;
+      if (xla != null)
+        xla.CloseXL();
+      if (sh != null && connection != null)
+      {
+        sh.DeleteTable(connection, "products");
+        sh.CloseSQLConnection(conn: ref connection);
+      }
     }
 
+    /// <summary>
+    /// Calls the scrape function and adds it to the DB
+    /// </summary>
+    /// <param name="label">label in the UI to notify user if it was successful/failure</param>
+    /// <param name="mapsWeb">Map of product names and their URLs</param>
+    /// <param name="sh">SQLHandler Object inorder to close</param>
+    /// <param name="connection">to delete the table and disconnect</param>
     public void AddToDB(Dictionary<String, String> mapsWeb, System.Windows.Forms.Label label, SQLHandler sh, ref SqlConnection connection)
     {
+      //start a process to call python script
       System.Diagnostics.Process process = null;
       sh.OpenSQLConnection(conn: ref connection);
       sh.CreateTable(connection, "products", "name varchar(500), price int, review int, type varchar(500)");
 
       if (process == null)
       {
+        //add items to db
         process = new System.Diagnostics.Process();
         foreach (String mw in mapsWeb.Keys.ToList())
         {
@@ -120,14 +173,18 @@ namespace demo
             count++;
             if (count == 5)
             {
-              Error(label, "Unable to generate SQL");
+              Error(label, "Unable to generate SQL", null, sh, connection);
               break;
             }
             if (process.HasExited)
             {
               if (!File.Exists(outputFile))
-                Error(label, "Unable to generate SQL");
-              
+              {
+                Error(label, "Unable to generate SQL", null, sh, connection);
+                return;
+              }
+
+              //parse the xml and add to sql db
               XmlDocument data = new XmlDocument();
               data.Load(outputFile);
